small fixes

+++ src_man/algorithms.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   algorithms.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cmeng <cmeng@student.42.fr>                +#+  +:+       +#+        */
+/*   By: christianmeng <christianmeng@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/28 22:53:06 by cmeng             #+#    #+#             */
-/*   Updated: 2023/03/28 22:58:11 by cmeng            ###   ########.fr       */
+/*   Updated: 2023/04/01 17:29:41 by christianme      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	radix_sort(t_circle *stack_a, t_circle *stack_b, size_t size)
 	size_t	i;
 
 	pos = 0;
-	while ((1 << pos) <= stack_a->max_size)
+	while ((1 << pos) <= (int) stack_a->max_size)
 	{
 		i = 0;
 		if (!same_digits(size, stack_a, pos))
